Implemented a lot of services

+++ common-lib/src/main/kotlin/database/table/TeacherTable.kt
@@ -3,5 +3,5 @@ package com.github.heheteam.commonlib.database.tables
 import org.jetbrains.exposed.dao.id.IntIdTable
 
 object TeacherTable : IntIdTable("student") {
-  val tgId = integer("tgId")
+  val tgId = long("tgId")
 }
